requirements.txt
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
blis==0.7.4
catalogue==1.0.0
certifi==2020.12.5
chardet==4.0.0
coverage==4.5.1
coveralls==1.3.0
cymem==2.0.5
Django==3.1.4
djangorestframework==3.12.2
docopt==0.6.2
duckling==1.8.0
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-2.3.1/en_core_web_md-2.3.1.tar.gz
httpie==2.3.0
idna==2.10
JPype1==1.2.0
more-itertools==8.6.0
murmurhash==1.0.5
numpy==1.19.4
plac==1.1.3
pluggy==0.13.1
preshed==3.0.5
py==1.10.0
Pygments==2.7.3
PySocks==1.7.1
pytest==3.7.1
pytest-cov==2.5.1
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.1
requests-toolbelt==0.9.1
six==1.15.0
spacy==2.3.5
spacy-lookups-data==0.3.2
sqlparse==0.4.1
srsly==1.0.5
thinc==7.4.5
timefhuman==0.0.5
tqdm==4.54.1
urllib3==1.26.2
wasabi==0.8.0
